refactor: onSubmitHandler 로직을 component 밖으로 분리
